refactoring tflight and tfserving to use float32 instead of uint to fix tflight failed to set tensor error

+++ python/mnist-ensemble/model/pipeline_invoke_python.py
@@ -68,7 +68,7 @@ def invoke(request):
 def _transform_request(request):
     request_str = request.decode('utf-8')
     request_json = json.loads(request_str)
-    request_np = ((255 - np.array(request_json['image'], dtype=np.uint8)) / 255.0).reshape(1, 28, 28)
+    request_np = ((255 - np.array(request_json['image'], dtype=np.float32)) / 255.0).reshape(1, 28, 28)
     return {"image": request_np}
 
 # Note:  Don't change this...
